a new base class for MiniHack envs

+++ nle/env/minihack.py
@@ -1,32 +1,38 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 
 from nle.env.tasks import NetHackStaircase
-from nle.nethack import CompassDirection
-
+from nle.nethack import CompassDirection, NETHACKOPTIONS
+from nle.env.base import FULL_ACTIONS, NLE_SPACE_ITEMS
 
 import subprocess
 import os
+import gym
+
+# import numpy as np
 from shutil import copyfile
 
 PATH_DAT_DIR = os.path.join(os.path.dirname(os.path.realpath(__file__)), "dat")
 MOVE_ACTIONS = tuple(CompassDirection)
+# APPLY_ACTIONS = tuple(list(MOVE_ACTIONS) + [Command.PICKUP, Command.APPLY])
 
 
-def patch_nhdat(level_des):
+def patch_nhdat(des_file):
     fname = "./mylevel.des"
-    try:
-        with open(fname, "w") as f:
-            f.writelines(level_des)
-        _ = subprocess.call("nle/scripts/patch_nhdat.sh")
-    except Exception as e:
-        print("Something went wrong at level generation", e.args[0])
-    finally:
-        os.remove(fname)
-
-
-def patch_nhdat_existing(des_name):
-    try:
-        des_path = os.path.join(PATH_DAT_DIR, des_name)
+    if not des_file.endswith(".des"):
+        # If the des-file is passed as a string
+        try:
+            with open(fname, "w") as f:
+                f.writelines(des_file)
+            _ = subprocess.call("nle/scripts/patch_nhdat.sh")
+        except Exception as e:
+            print("Something went wrong at level generation", e.args[0])
+        finally:
+            os.remove(fname)
+    else:
+        # Use the .des file if exists, otherwise search in minihack directory
+        des_path = os.path.abspath(des_file)
+        if not os.path.exists(des_path):
+            des_path = os.path.join(PATH_DAT_DIR, des_file)
         if not os.path.exists(des_path):
             print(
                 "{} file doesn't exist. Please provide a path to a valid .des \
@@ -34,39 +40,113 @@ def patch_nhdat_existing(des_name):
                     des_path
                 )
             )
-        fname = "./mylevel.des"
-        copyfile(des_path, fname)
-        _ = subprocess.call("nle/scripts/patch_nhdat.sh")
-    except Exception as e:
-        print("Something went wrong at level generation", e.args[0])
-    finally:
-        os.remove(fname)
+        try:
+            copyfile(des_path, fname)
+            _ = subprocess.call("nle/scripts/patch_nhdat.sh")
+        except Exception as e:
+            print("Something went wrong at level generation", e.args[0])
+        finally:
+            os.remove(fname)
 
 
-class MiniHackMaze(NetHackStaircase):
-    """Base environment for maze-type task. """
+class MiniHackCustom(NetHackStaircase):
+    """Base class for custom MiniHack environments.
+
+    Features:
+    - Default nethack options
+    - Full action space by default
+    - Wizard mode is turned off by default
+    - One-letter menu questions are allowed by default
+    - Includes all NLE observations
+
+    The goal is to reach the staircase.
+
+    Use cases:
+    - Use this class if you want to experiment with different description files
+    and require rich (full) action space.
+    - Use a MiniHackMaze class for maze-type environments where there is no pet,
+    action space is severely restricted and no one-letter questions are required.
+    - Inherit from this class if you require a different reward function and
+    dynamics. You might need to override the following methods
+        - self._is_episode_end()
+        - self._reward_fn()
+        - self.step()
+        - self.reset()
+    """
 
     def __init__(self, *args, des_file: str = None, **kwargs):
         # No pet
-        kwargs["options"] = kwargs.pop("options", [])
-        kwargs["options"].append("pettype:none")
+        kwargs["options"] = kwargs.pop("options", list(NETHACKOPTIONS))
         # Actions space - move only
-        kwargs["actions"] = MOVE_ACTIONS
-        # Enter Wizard mode
-        kwargs["wizard"] = kwargs.pop("wizard", True)
-        # Override episode limit
-        kwargs["max_episode_steps"] = kwargs.pop("max_episode_steps", 100)
+        kwargs["actions"] = kwargs.pop("actions", FULL_ACTIONS)
+        # Enter Wizard mode - turned off by default
+        kwargs["wizard"] = kwargs.pop("wizard", False)
+        # Allowing one-letter menu questions
+        kwargs["allow_all_yn_questions"] = kwargs.pop("allow_all_yn_questions", True)
+        # Episode limit
+        kwargs["max_episode_steps"] = kwargs.pop("max_episode_steps", 200)
+        # Using all NLE observations by default
+        space_dict = dict(NLE_SPACE_ITEMS)
+        # Not currently passing the observation keys to the base class
+        # because they are used in render(), which is used when developing
+        # new environments. Instead, we filter the observations in the
+        # _get_observation() method we override.
+        self._minihack_obs_keys = kwargs.pop(
+            "observation_keys", list(space_dict.keys())
+        )
 
         # Patch the nhddat library by compling the given .des file
         if des_file is None:
-            raise ValueError("Description filename is not provided.")
+            raise ValueError("Description file is not provided.")
 
-        if des_file.endswith(".des"):
-            patch_nhdat_existing(des_file)
-        else:
-            patch_nhdat_existing(des_file)
+        patch_nhdat(des_file)
 
         super().__init__(*args, **kwargs)
+
+        self.observation_space = gym.spaces.Dict(
+            {key: space_dict[key] for key in self._minihack_obs_keys}
+        )
+
+    def _get_observation(self, observation):
+        # Filter out observations that we don't need
+        observation = super()._get_observation(observation)
+        return {
+            key: val
+            for key, val in observation.items()
+            if key in self._minihack_obs_keys
+        }
+
+
+class MiniHackMaze(MiniHackCustom):
+    """Base class for maze-type task.
+
+    Maze environments have
+    - Restricted action space (move only by default)
+    - No pet
+    - One-letter menu questions are NOT allowed by default
+    - Restricted observations, only glyphs by default
+    - No random monster generation
+
+    The goal is to reach the staircase.
+    """
+
+    def __init__(self, *args, des_file: str = None, **kwargs):
+        # No pet
+        kwargs["options"] = kwargs.pop("options", list(NETHACKOPTIONS))
+        kwargs["options"].append("pettype:none")
+        # No random monster generation after every timestep
+        # As a workaround to a current issue, we are utilizing the nudist option instead
+        kwargs["options"].append("nudist")
+        # Actions space - move only
+        kwargs["actions"] = kwargs.pop("actions", MOVE_ACTIONS)
+        # Disallowing one-letter menu questions
+        kwargs["allow_all_yn_questions"] = kwargs.pop("allow_all_yn_questions", False)
+        # Override episode limit
+        kwargs["max_episode_steps"] = kwargs.pop("max_episode_steps", 100)
+        # Restrict the observation space to glyphs only
+        kwargs["observation_keys"] = kwargs.pop("observation_keys", ["glyphs"])
+
+        super().__init__(*args, des_file=des_file, **kwargs)
 
 
 class MiniHackEmpty(MiniHackMaze):
